Split up the facade with service classes and repository for separation of concerns
